docker-hands-on/demo-app/pyproject.toml
[project]
name = "demo-app"
version = "0.1.0"
description = "Demo Application to introduce Docker concepts"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
